requirements.txt
fastapi~=0.116.1
dotenv~=0.9.9
python-dotenv~=1.1.1
python-telegram-bot==22.3
SQLAlchemy==2.0.41
mysql-connector-python==9.4.0
langchain==0.3.27
langchain-community==0.3.27
openai~=1.98.0